treehouse-project-1
In this project, you'll create an app that displays random famous
quotes each time a button is clicked. You can display a quote from a
famous athlete, politician, or historical figure:

"The only thing we have to fear is fear itself." — Franklin Delano
Roosevelt.

You'll need to use your knowledge of basic JavaScript syntax, including
variables, loops and object literals, to complete this project.

To help you get started, we'll give you basic HTML and CSS, as well a
JavaScript file containing some starter code. But you'll have to find
the quotes yourself and build the data structure to store them in.

This project is a fun and effective way for you to practice fundamental
JavaScript skills. It also gives you a simple interactive portfolio
piece to show off your understanding of JavaScript.